Create IMAGE_INSCRIBOOR_V1
This is a python script that is meant to be utilized in PowerShell. 
Store this file in the same directory as your doginals.js file. 
Before executing this script, open it in PowerShell and adjust the parameters in the variables: 

directory = 'C:\\Users\\YourMom\\Desktop\\NodeRunners\\Hi-res'
# replace with file extension
file_extension = 'png'
# enter range of files to inscribe.
start = 8
end = 33


This script will look inside of the directory for files named "[start number].png to [end number].png within the directory that you set. 

You can change the file type within the file_extension variable, which is currently set to 'png'